% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append}
\alias{append}
\title{Vector Merging}
\usage{
append(x, values, after = length(x))
}
\arguments{
\item{x}{the vector to be modified.}

\item{values}{to be included in the modified vector.}

\item{after}{a subscript, after which the values are to be appended.}
}
\value{
If \code{x} and \code{values} are of class \code{factor}, 
  the function returns an object of class \code{factor} whereas levels are 
  determined using \code{\link{cfactor}} internally. Levels are only
  preserved if both elements have the same levels. In addition, ordering
  is only preserved if the order is also identical.
  Otherwise, the base 
  method is called directly.
}
\description{
Vector Merging
}
\details{
A refined version of \code{\link[base]{append}} that allows for
  factor appending.
}
\examples{
f <- factor(c('c','a','a',NA,'b','a'), levels= c('a','b','c'))
g <- factor(sample(letters[4:10]), levels = sample(letters[4:10]))
base::append(f, g) # not nice
refactor::append(f, g) # probably what you want
}

