% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{c.factor}
\alias{c.factor}
\title{Combine Values into a Factor}
\usage{
\method{c}{factor}(..., recursive = FALSE)
}
\arguments{
\item{...}{objects to be concatenated.}

\item{recursive}{logical.  If \code{recursive = TRUE}, the function
    recursively descends through lists (and pairlists) combining all
    their elements into a vector.}
}
\description{
Combine Values into a Factor
}
\examples{
char1 <- as.factor(sample(letters, 5))
char2 <- as.factor(sample(letters, 5))
c(char1, char2)
}

