% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.R
\name{c.ordered}
\alias{c.ordered}
\title{Combine Values into an Ordered Factor}
\usage{
\method{c}{ordered}(..., recursive = FALSE)
}
\arguments{
\item{...}{objects to be concatenated.}

\item{recursive}{logical.  If \code{recursive = TRUE}, the function
    recursively descends through lists (and pairlists) combining all
    their elements into a vector.}
}
\description{
Combine Values into an Ordered Factor
}
\examples{
group1 <- cfactor(c("US$51 - US$75", "US$101 - US$125", "US$1 - US$25"), 
                 ordered = TRUE)
group2 <- cfactor(c("US$76 - US$100", "US$26 - US$50", "US$126 - US$150"), 
                 ordered = TRUE)
c(group1, group2)
}

