% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc.R
\name{cc}
\alias{cc}
\title{Combine Values into an (ordered) Factor}
\usage{
cc(..., recursive = FALSE)
}
\arguments{
\item{...}{objects to be concatenated.}

\item{recursive}{logical.  If \code{recursive = TRUE}, the function
    recursively descends through lists (and pairlists) combining all
    their elements into a vector.}
}
\description{
Combine Values into an (ordered) Factor
}
\details{
Note that the order of the levels is internally determined by
\code{\link{cfactor}} and thus its ordering scheme. Refer to the section 
'Details' in the \code{\link{cfactor}} documentation.
}
\examples{
char1 <- as.factor(sample(letters, 5))
char2 <- as.factor(sample(letters, 5))
cc(char1, char2)
}

