% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_files_parse.R
\name{check_files_parse}
\alias{check_files_parse}
\title{Check that files parse correctly}
\usage{
check_files_parse(path = ".", recursive = TRUE)
}
\arguments{
\item{path}{A string. The path to a file or the folder to explore
By default explores the working directory.}

\item{recursive}{A boolean. Passed to \code{list.files()} if \code{path} is a directory}
}
\value{
Returns the path invisibly, called for side effects.
}
\description{
This identifies files that contain non syntactic code, including files that
have an R extension despite not being an R script.
}
