% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_similar_code.R
\name{detect_similar_code}
\alias{detect_similar_code}
\title{Detect similar code blocks}
\usage{
detect_similar_code(paths = ".", recursive = TRUE, pattern = NULL)
}
\arguments{
\item{paths}{Paths to scripts or folders containing scripts}

\item{recursive}{Whether folders should be explorer recursively}

\item{pattern}{A regular expression used to filter files}
}
\value{
The \code{paths} argument invisibly. Called for side effects
}
\description{
This is a wrapper around \code{dupree::dupree()}. It analyses provided files looking for
similar code. It is a bit slow so it can be impractical on big projects to run
it with default (all files contained in working directory recursively), in
this case it is wiser to run it on one or more specific \code{paths} or/and
or to filter the files using the \code{pattern} argument.
}
