% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_package_funs.R
\name{find_pkg_funs}
\alias{find_pkg_funs}
\title{Find all uses of a package's functions}
\usage{
find_pkg_funs(pkg, path = ".", recursive = TRUE, exclude = NULL)
}
\arguments{
\item{pkg}{A string. The name of the target package}

\item{path}{A string. The path to a file or the folder to explore
By default explores the working directory.}

\item{recursive}{A boolean. Passed to \code{list.files()} if \code{path} is a directory}

\item{exclude}{A character vector of function names to dismiss.}
}
\value{
Returns its input invisibly, called for side effects
}
\description{
This will show false positives, but guarantees that we don't miss any instance.
}
