% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_hybrid_scripts.R
\name{identify_hybrid_scripts}
\alias{identify_hybrid_scripts}
\title{Identify hybrid scripts}
\usage{
identify_hybrid_scripts(path = ".", recursive = TRUE)
}
\arguments{
\item{path}{A string. The path to a file or the folder to explore
By default explores the working directory.}

\item{recursive}{A boolean. Passed to \code{list.files()} if \code{path} is a directory}
}
\value{
Returns the path invisibly, called for side effects.
}
\description{
Identify scripts who contain both function definitions and other object definitions.
}
