% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_cfactor.R
\name{index_cfactor}
\alias{index_cfactor}
\title{index_cfactor}
\usage{
index_cfactor(data, index, variable = "variable", encoding = "encoding",
  label = "label", ...)
}
\arguments{
\item{data}{A data.frame containing integers to decode.}

\item{index}{A data.frame containing the encoding for \code{data}.}

\item{variable}{The name of the column in \code{index} that indicates the variable.}

\item{encoding}{The name of the column in \code{index} that indicates the encoding.}

\item{...}{Further arguments to pass to \code{\link{cfactor}}.}
}
\description{
Decode numerical data into (ordered) factors given the encoding
}
\examples{
data <- data.frame(var1 = sample(x = 1:10, size = 20, replace = TRUE),
                   var2 = rep(1:2, 20),
                   var3 = sample(20),
                   var4 = 2, 
                   var5 = sample(row.names(USArrests), size = 20),
                   stringsAsFactors = FALSE)

index <- data.frame(var = rep(paste0("var", 1:3), c(10, 2, 20)),
                   encoding = c(1:10, 1:2, 1:20),
                   label = c(letters[1:10], c("male", "female"), LETTERS[1:20]))
                   
index_cfactor(data = data, index = index, variable = "var", ordered = c(TRUE, TRUE, FALSE))
}

