% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor-package.R
\docType{package}
\name{refactor-package}
\alias{refactor}
\alias{refactor-package}
\title{refactor: Tools for Refactoring Code}
\description{
Tools for refactoring code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moodymudskipper/refactor}
  \item Report bugs at \url{https://github.com/moodymudskipper/refactor/issues}
}

}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@gmail.com}

}
\keyword{internal}
