% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_function.R
\name{rename_function}
\alias{rename_function}
\title{Rename a function}
\usage{
rename_function(name, new_name, file_name, language = "R",
  calls = TRUE, defs = TRUE, examples = TRUE, characters = TRUE,
  comments = TRUE, docs = TRUE, verbosity_level = "normal",
  idle_mode = FALSE)
}
\arguments{
\item{name}{`character` string giving a function name to be renamed}

\item{new_name}{`character` string giving a new name of the function}

\item{language}{`character` string giving a language in which file is written. Currently
supported language is R}

\item{calls}{`logical` value. Should the function be renamed in calls?}

\item{defs}{`logical` value. Should the function be renamed in definitions?}

\item{examples}{`logical` value. Should the function be renamed in examples in documentation?}

\item{characters}{`logical` value. Should the function be renamed in literal character vectors?
This option do not make difference between the function and other objects!}

\item{comments}{`logical` value. Should function be renamed in comments?
This option do not make difference between the function and other objects!}

\item{docs}{`logical` value. Should function be renamed in text in documentation?
This option do not make difference between the function and other objects!}

\item{verbosity_level}{`logical` value. Allows changing amount of the information printed in the
console.}

\item{idle_mode}{`logical` value. If `TRUE`, changes will not be applied, only information will
be printed.}
}
\description{
Renames a function in a file
}
\author{
Dominik Rafacz
}
