% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lintr_wrappers.R
\name{use_lintr_template_on_file}
\alias{use_lintr_template_on_file}
\alias{use_lintr_template_on_dir}
\title{Use lintr template}
\usage{
use_lintr_template_on_file(path = NULL)

use_lintr_template_on_dir(path = NULL)
}
\arguments{
\item{path}{Path to a R script or a directory. By default \code{use_lint_template_on_file()}
considers the active document and \code{use_lint_template_on_dir()} considers the
project folder as returned by \code{here::here()}}
}
\value{
Returns \code{NULL} invisibly. Called for side effects.
}
\description{
This opens up an untitled script in RStudio containing calls to \code{lintr::lint()}
or \code{lintr::lint_dir()} with various linters, sorted by category and rough
order of importance.
}
