% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_namespace_names.R
\name{use_namespace_check}
\alias{use_namespace_check}
\title{Use namespace check}
\usage{
use_namespace_check()
}
\value{
Returns \code{NULL} invisibly, called for side effects.
}
\description{
Wrapper around \code{fetch_namespace_names()} that opens a new RStudio source editot
tab with code to be pasted at the top of the main script to enumerate required
packages and test if they are installed.
}
