% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfactor.R
\name{cfactor}
\alias{cfactor}
\title{Factor}
\usage{
cfactor(x, levels = NULL, labels = levels, exclude = NA,
  ordered = is.ordered(x), nmax = NA)
}
\arguments{
\item{x}{A vector of data, usually taking a small number of distinct values.}

\item{levels}{An optional vector of the values (as character strings) that x might have taken. The default is the unique set
of values taken by as.character(x), sorted into increasing order of x. Note that this set can be specified as smaller than
sort(unique(x)).}

\item{labels}{Either an optional character vector of labels for the levels (in the same order as levels after removing those in exclude),
or a character string of length 1.}
}
\value{
An object of class "factor" which has a set of integer codes the length of x with a "levels" attribute of mode character
and unique (!anyDuplicated(.)) entries. If argument ordered is true (or ordered() is used) the result has class c("ordered", "factor").
}
\description{
A wrapper for factor with enhanced control.
}
\examples{
cfactor(c("a", "c", "b", "c", "d"))
}

