% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.ordered.R
\name{cut.ordered}
\alias{cut.ordered}
\title{Create Bins for ordered Factors}
\usage{
\method{cut}{ordered}(x, breaks, labels = NULL, include.lowest = TRUE,
  right = TRUE, ordered_result = TRUE, label_sep = "-", ...)
}
\arguments{
\item{x}{An ordered factor which is to be cut into ordered bins.}

\item{breaks}{Either a numeric vector of two or more unique cut points or a 
 single number (greater than or equal to 2) giving the
number of intervals into which \code{x} is to be cut.}

\item{labels}{Labels for the levels of the resulting category. By default, 
labels are constructed using \code{a-b, c-d} interval notation.
If \code{labels = FALSE}, simple integer codes are returned instead of a 
factor.}

\item{include.lowest}{Logical, indicating if an \code{x[i]} equal to the 
lowest (or highest, for \code{right = FALSE}) breaks value should be 
included.}

\item{right}{Logical, indicating how to create the bins. This is utilized in
two different ways based on the type of breaks argument. In the
conventional case, where a breaks vector is supplied, \code{right = TRUE}
indicates that bins should be closed on the right (and open on the left) or
vice versa. If a single integer breaks value is provided, then 
\code{right = TRUE} indicates that bins will be determined such that those 
on the right are larger (if it is not possible for all bins to be evenly 
sized).}

\item{ordered_result}{Logical: should the result be an ordered factor? Note 
that since the input data is ordered this argument is \code{TRUE} by 
default.}

\item{label_sep}{A single or short character string used to generate labels 
for the intervals e.g. the default value of "-" will result in labels like 
a-c d-g i-z etc.}

\item{...}{Further arguments to be passed to or from other methods, 
in particular to \code{\link{cut.default}}.}
}
\value{
A factor is returned, unless \code{labels = FALSE} which results in 
 an integer vector of level codes.
}
\description{
cut divides the range of \code{x} into intervals and codes the values in 
 \code{x} according to the interval they fall into.
}
\details{
Note that the \code{dig.lab} argument from \code{cut.default}, is 
 replaced for \code{cut.ordered} by a new argument: \code{label_sep}. Also 
 note that unlike \code{\link[base]{cut.default}}, here \code{include.lowest} 
 defaults to \code{TRUE}, since this is more intuitive for the class 
 \code{ordered}. Finally, since the input for \code{cut.ordered} is of class 
 \code{ordered}, this is the default output as well. This contrasts with 
 \code{cut.default}, which produces an unordered output by default.
}
\examples{
 some_letters <- cfactor(letters, ordered = TRUE)
 head(cut(some_letters, breaks = c("a", "q", "z"), 
          labels = c("beginning of the alphabet", 
                     "the rest of the alphabet"), 
          right = TRUE, include.lowest = TRUE))
}

