% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.ordered.R
\name{cut.ordered}
\alias{cut.ordered}
\title{Convert Numeric to Factor}
\usage{
\method{cut}{ordered}(x, breaks, labels = NULL, include.lowest = FALSE,
  right = TRUE, digit.lab = 3, ordered_result = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector which is to be converted to a factor by cutting.}

\item{breaks}{Either a numeric vector of two or more unique cut points or a single number (greater than or equal to 2) giving the
number of intervals into which x is to be cut.}

\item{labels}{Labels for the levels of the resulting category. By default, labels are constructed using \code{(a,b]} interval notation.
If \code{labels = FALSE}, simple integer codes are returned instead of a factor.}

\item{include.lowest}{Logical, indicating if an \code{x[i]} equal to the lowest (or highest, for \code{right = FALSE}) breaks value 
should be included.}

\item{right}{Logical, indicating if the intervals should be closed on the right (and open on the left) or vice versa.}

\item{digit.lab}{Integer which is used when labels are not given. It determines the number of digits used in formatting the break
numbers.}

\item{ordered_result}{Logical: should the result be an ordered factor?}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A factor is returned, unless labels = FALSE which results in an integer vector of level codes.
}
\description{
cut divides the range of x into intervals and codes the values in x according to the interval they fall into.
}
\examples{
Z <- stats::rnorm(10000)
cut(Z, breaks = -6:6)
}

