% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor.R
\name{refactor}
\alias{refactor}
\title{Factors, Redux}
\usage{
refactor(x, levels = NULL, logic = c("freq", "appear"), ...)
}
\arguments{
\item{x}{A vector of data.}

\item{levels}{An optional character vector to be used as the levels. If
specified, \code{x}, \code{levels}, and additional arguments via \code{...}
are passed through to \code{\link{factor}}.}

\item{logic}{The logical principle by which levels of \code{x} should be set.
If \code{freq} (the default), the levels are in order of decreasing
frequency, with ties broken alphabetically. If \code{appear}, the levels
are in order of appearance.}

\item{...}{Additional arguments for \code{base::factor()}, e.g.
\code{ordered}.}
}
\value{
A factor.
}
\description{
Factors, Redux
}
\examples{
set.seed(1)
x <- sample(letters, size = 20, replace = TRUE)

if (requireNamespace("ggplot2")) {
  library(ggplot2)

  ## refactor on-the-fly to get ggplot2 barplot sorted by frequency
  ggplot(data.frame(x = refactor(x)), aes(x)) + geom_bar()

  ## or create the factor explicitly
  xfact <- refactor(x)
  ggplot(data.frame(x = xfact), aes(x)) + geom_bar()

  ## set levels in order of appearance in the data
  xfact_appear <- refactor(x, logic = "appear")
  ggplot(data.frame(x = xfact_appear), aes(x)) + geom_bar()
}
}
\seealso{
\code{\link{factor}}
}

