% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor.R
\docType{package}
\name{refactor}
\alias{refactor}
\alias{refactor-package}
\title{refactor: Better factor handling for R}
\description{
refactor aims to provide better handling of factors by addressing some of the known problems and 
short-comings with the existing toolset for working with factors in R.
}

