% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor.r
\name{refactor}
\alias{refactor}
\alias{\%refactor\%}
\alias{\%refactor2\%}
\title{Refactor Code}
\usage{
original \%refactor\% refactored

original \%refactor2\% refactored
}
\arguments{
\item{original}{original expression}

\item{refactored}{refactored expression}
}
\description{
Both original and refactored expressions are run. The function will fail if
the outputs are different. \verb{\%refactor2\%} will also fail if the refactored
expression runs slower.
}
\examples{
{
apply(cars, 1, max)
} \%refactor\% {
  do.call(pmax, cars)
}
}
